add more comments and code clean-up
