+++ src/pages/groupComparison/GroupComparisonStore.ts
@@ -47,7 +47,6 @@ import {
 import { fetchPatients } from 'pages/resultsView/ResultsViewPageStoreUtils';
 import { isSampleProfiled } from 'shared/lib/isSampleProfiled';
 import { getSampleMolecularIdentifiers } from 'pages/studyView/StudyViewComparisonUtils';
-import { LEFT_TRUNCATION_ADJUSTMENT_FLAG } from 'pages/resultsView/survival/SurvivalUtil';
 import { FeatureFlagEnum } from 'shared/featureFlags';
 
 export default class GroupComparisonStore extends ComparisonStore {
@@ -623,7 +622,7 @@ export default class GroupComparisonStore extends ComparisonStore {
     }
 
     // override parent method
-    protected get isLeftTruncationForSurvivalDataEnabled() {
+    protected get isLeftTruncationFeatureFlagEnabled() {
         return this.appStore.featureFlagStore.has(
             FeatureFlagEnum.LEFT_TRUNCATION_ADJUSTMENT
         );
