+++ src/pages/groupComparison/styles.module.scss
@@ -167,7 +167,7 @@
     }
 }
 
-.NoPaddingLeftTruncationCheckbox {
+.noPaddingLeftTruncationCheckbox {
     background-color: white;
     margin-bottom: 3px;
 }
