+++ src/pages/resultsView/SurvivalStoreHelper.ts
@@ -27,6 +27,8 @@ export function getPatientSurvivals(
                     const monthsClinicalData = clinicalData.find(
                         c => c.clinicalAttributeId === monthsAttributeId
                     );
+                    // if entry months data is not available, we can assume all data is collected at the beginning
+                    // So we can assume entryMonths is 0 by default
                     const entryMonths =
                         entryMonthsByUniquePatientKey?.[uniquePatientKey] || 0;
                     if (
