+++ src/pages/resultsView/survival/styles.module.scss
@@ -39,7 +39,7 @@
     display: inline-block;
 }
 
-.PaddingLeftTruncationCheckbox {
+.paddingLeftTruncationCheckbox {
     padding: 0px 10px;
     background-color: white;
     margin-bottom: 3px;
