+++ src/pages/resultsView/survival/styles.module.scss.d.ts
@@ -1,9 +1,9 @@
 declare const styles: {
   readonly "NotAvailable": string;
-  readonly "PaddingLeftTruncationCheckbox": string;
   readonly "SurvivalTable": string;
   readonly "Tooltip": string;
   readonly "XmaxNumberInput": string;
+  readonly "paddingLeftTruncationCheckbox": string;
 };
 export = styles;
